\name{IsingFit-package}
\alias{IsingFit-package}
\docType{package}
\title{
Network estimation using the eLasso method
}
\description{
This network estimation procedure eLasso, which is based on the Ising model, combines l1-regularized logistic regression with model selection based on the Extended Bayesian Information Criterion (EBIC). EBIC is a fit measure that identifies relevant relationships between variables. The resulting network consists of variables as nodes and relevant relationships as edges.
}
\details{
\tabular{ll}{
Package: \tab IsingFit\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-22\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Claudia D. van Borkulo, Sacha Epskamp

Maintainer: Claudia D. van Borkulo <cvborkulo@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
Chen, J., & Chen, Z. (2008). Extended bayesian information criteria for model selection with large model spaces. Biometrika, 95(3), 759-771.
Foygel, R., & Drton, M. (2011). Bayesian model choice and information criteria in sparse generalized linear models. arXiv preprint arXiv:1112.5635.
Van Borkulo, C. D., Borsboom, D., Epskamp, S., Blanken, T. F., Boschloo, L., Schoevers, R. A., & Waldorp, L. J. (manuscript submitted for publication). Fitting like a glove: New methodology for constructing networks from psychometric data.
}
% ~~ Optionally other standard keywords, one per line, from file ~~
% ~~ KEYWORDS in the R documentation directory ~~

